Removed Typescript runtime.
