+++ packages/cli/package.json
@@ -20,24 +20,13 @@
     "listr-update-renderer": "^0.4.0",
     "listr-verbose-renderer": "^0.4.1",
     "lodash": "^4.17.11",
-    "tslib": "^1",
     "yeoman-environment": "^2.3.3",
     "yeoman-generator": "^3.1.1"
   },
   "devDependencies": {
     "@oclif/dev-cli": "1.19.0",
     "@oclif/test": "1.2.1",
-    "@oclif/tslint": "3.1.0",
-    "@types/chai": "4.1.6",
-    "@types/mocha": "5.2.5",
-    "@types/node": "10.11.6",
-    "chai": "4.2.0",
-    "globby": "8.0.1",
-    "mocha": "5.2.0",
-    "nyc": "13.0.1",
-    "ts-node": "7.0.1",
-    "tslint": "5.11.0",
-    "typescript": "3.1.2"
+    "globby": "8.0.1"
   },
   "engines": {
     "node": ">=8.0.0"
@@ -55,7 +44,7 @@
   "license": "Apache-2.0",
   "main": "lib/index.js",
   "oclif": {
-    "commands": "./lib/commands",
+    "commands": "./src/commands",
     "bin": "wcf",
     "plugins": [
       "@oclif/plugin-help"
@@ -64,9 +53,8 @@
   "repository": "elmsln/wcfactory",
   "scripts": {
     "postpack": "rm -f oclif.manifest.json npm-shrinkwrap.json",
-    "posttest": "tslint -p test -t stylish",
     "prepack": "oclif-dev manifest && oclif-dev readme && npm shrinkwrap",
-    "test": "nyc --extension .ts mocha --forbid-only \"test/**/*.test.ts\"",
+    "test": "echo NO TESTS",
     "version": "oclif-dev readme && git add README.md"
   },
   "types": "lib/index.d.ts",
