--- packages/cli/src/index.ts
@@ -1,1 +1,0 @@
-export {run} from '@oclif/command'
