--- packages/cli/test/commands/build.test.ts
@@ -1,17 +1,0 @@
-import {expect, test} from '@oclif/test'
-
-describe('build', () => {
-  test
-    .stdout()
-    .command(['build'])
-    .it('runs hello', ctx => {
-      expect(ctx.stdout).to.contain('hello world')
-    })
-
-  test
-    .stdout()
-    .command(['build', '--name', 'jeff'])
-    .it('runs hello --name jeff', ctx => {
-      expect(ctx.stdout).to.contain('hello jeff')
-    })
-})
