--- packages/cli/test/commands/element.test.ts
@@ -1,17 +1,0 @@
-import {expect, test} from '@oclif/test'
-
-describe('element', () => {
-  test
-    .stdout()
-    .command(['element'])
-    .it('runs hello', ctx => {
-      expect(ctx.stdout).to.contain('hello world')
-    })
-
-  test
-    .stdout()
-    .command(['element', '--name', 'jeff'])
-    .it('runs hello --name jeff', ctx => {
-      expect(ctx.stdout).to.contain('hello jeff')
-    })
-})
