--- packages/cli/test/commands/factory.test.ts
@@ -1,17 +1,0 @@
-import {expect, test} from '@oclif/test'
-
-describe('init', () => {
-  test
-    .stdout()
-    .command(['init'])
-    .it('runs hello', ctx => {
-      expect(ctx.stdout).to.contain('hello world')
-    })
-
-  test
-    .stdout()
-    .command(['init', '--name', 'jeff'])
-    .it('runs hello --name jeff', ctx => {
-      expect(ctx.stdout).to.contain('hello jeff')
-    })
-})
