--- packages/cli/test/commands/start.test.ts
@@ -1,17 +1,0 @@
-import {expect, test} from '@oclif/test'
-
-describe('start', () => {
-  test
-    .stdout()
-    .command(['start'])
-    .it('runs hello', ctx => {
-      expect(ctx.stdout).to.contain('hello world')
-    })
-
-  test
-    .stdout()
-    .command(['start', '--name', 'jeff'])
-    .it('runs hello --name jeff', ctx => {
-      expect(ctx.stdout).to.contain('hello jeff')
-    })
-})
