--- packages/cli/test/mocha.opts
@@ -1,5 +1,0 @@
---require ts-node/register
---watch-extensions ts
---recursive
---reporter spec
---timeout 5000
