--- packages/cli/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "importHelpers": true,
-    "module": "commonjs",
-    "outDir": "lib",
-    "rootDir": "src",
-    "strict": true,
-    "target": "es2015",
-    "composite": true
-  },
-  "include": [
-    "src/**/*"
-  ]
-}
