--- packages/cli/tslint.json
@@ -1,3 +1,0 @@
-{
-  "extends": "@oclif/tslint"
-}
